Moved file transfer receiving logic from getfile to fileTransferService, Improvements.
