+++ app/getfile.js
@@ -6,8 +6,10 @@ var fs = require('fs');
 
 var constants = require('./constants');
 var pairingService = require('./pairingService');
+var fileTransferService = require('./fileTransferService');
 
 var broadcastListener = new pairingService.BroadcastListener();
+var fileReceiverClient = new fileTransferService.FileReceiver();
 var thisIP = ip.address();
 var hostIP;
 var fileName;
@@ -21,7 +23,7 @@ broadcastListener.listenForBroadcast(handshakeMadeHandler, thisIP);
 function handshakeMadeHandler(receivedHostIP, receivedFileName){
   hostIP = receivedHostIP;
   fileName = receivedFileName;
-  setTimeout(beginTransfer, 2000);
+  setTimeout(transferFile, 4000);
 }
 
 
@@ -30,34 +32,10 @@ function handshakeMadeHandler(receivedHostIP, receivedFileName){
 //*************Connect to server, Initiate transfer  (TCP Socket)
 //----------------
 
-function beginTransfer() {
+function transferFile(){
+  fileReceiverClient.beginTransfer(hostIP, thisIP, fileName, transferFinishedHandler);
+}
 
-  var socketConnection = new netModule.Socket();
-
-  socketConnection.connect(constants.PORT, hostIP, function connected(){
-    //----------Config and handshake-------------------------------
-    //Kill Socket after 5 seconds inactivity
-    socketConnection.setTimeout(10000, function(){
-      console.log('Socket Timeout');
-      socketConnection.destroy();
-      process.exit(1);
-    });
-
-    socketConnection.write('Client at: '+ thisIP + ' confirms connection.');
-    
-
-    //-----------Receive data and write to file in current directory-------------
-    var writeStream = fs.createWriteStream(process.cwd() + '/' + fileName);
-    socketConnection.on('data', function(chunk){
-      writeStream.write(chunk);
-    });
-
-    socketConnection.on('end', function(){
-      console.log('File Received');
-      console.log('Closing Connection');
-      writeStream.end();
-      socketConnection.destroy();
-    });
-
-  });
-};
+function transferFinishedHandler(exitStatus){
+  process.exit(exitStatus);
+}
